[ONES-82456247] Optimize the plus publishing script to avoid the issue of gluten jars being overwritten during spark publishing

+++ plus/run.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 
-cd /opt/meituan/spark-3.0-vec/jars/
-rm gluten-package*SNAPSHOT.jar
-cp /opt/meituan/gluten-1.1.0/gluten-package*SNAPSHOT.jar /opt/meituan/spark-3.0-vec/jars/
-ls -l |grep gluten-package*
+if [ -d /opt/meituan/spark-3.0-vec/jars/ ]; then
+    cd /opt/meituan/spark-3.0-vec/jars/
+    rm gluten-package*SNAPSHOT.jar
+    cp /opt/meituan/gluten/target/gluten-package*SNAPSHOT.jar /opt/meituan/spark-3.0-vec/jars/
+    ls -l |grep gluten-package*
+else
+    echo "Dir /opt/meituan/spark-3.0-vec/jars/ is not exist, just continue"
+fi
